aod-touchable: Move all webext files to app assets
Turns out to install a builtin webextension an __unzipped__ extension
__folder__ is needed. Luckily when the apk is created symlinks seem to
be resolved beforehand.
This also makes the Makefile creating a xpi redundandt, so it is removed.

Because I'm too lazy to split it up properly this also contains a typo
fix for the extension id, and some commented experimentation regarding a
few containers that are still not touch-friendly.